03_sqlite_to_postgres/requirements.txt
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytest==7.1.1